mac error

+++ inst/fmalloc/sysdeps/pthread/malloc-machine.h
@@ -26,6 +26,7 @@ PERFORMANCE OF THIS SOFTWARE.
 #define _PTHREAD_MALLOC_MACHINE_H
 
 #include <pthread.h>
+#include <stdint.h>
 
 #undef thread_atfork_static
 
@@ -108,9 +109,9 @@ typedef void *tsd_key_t[256];
   for(i=0; i<256; i++) (*key)[i] = 0; \
 } while(0)
 #define tsd_setspecific(key, data) \
- (key[(unsigned)pthread_self() % 256] = (data))
+ (key[(uintptr_t)pthread_self() % 256] = (data))
 #define tsd_getspecific(key, vptr) \
- (vptr = key[(unsigned)pthread_self() % 256])
+ (vptr = key[(uintptr_t)pthread_self() % 256])
 
 #else
